MBS-10881 add settings convention plugins
for:
- build cache
- scan
- dependencyResolutionManagement
- pluginManagement
