+++ build-logic-settings/cache-plugin/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    `kotlin-dsl`
+}
