+++ build-logic-settings/dependency-plugin/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    `kotlin-dsl`
+}
+
+dependencies {
+    implementation("com.gradle.enterprise:com.gradle.enterprise.gradle.plugin:3.6.3")
+}
