+++ build-logic-settings/dependency-plugin/src/main/kotlin/com/avito/android/artifactory/ArtifactoryProxy.kt
@@ -1,0 +1,18 @@
+package com.avito.android.artifactory
+
+import org.gradle.api.artifacts.repositories.MavenArtifactRepository
+
+fun MavenArtifactRepository.setUrlOrProxy(artifactoryUrl: String?, repositoryName: String, originalRepo: String) {
+    if (artifactoryUrl.isNullOrBlank()) {
+        name = repositoryName
+        setUrl(originalRepo)
+    } else {
+        name = "Proxy for $repositoryName: $originalRepo"
+        artifactoryUrl(artifactoryUrl, repositoryName)
+    }
+}
+
+private fun MavenArtifactRepository.artifactoryUrl(artifactoryUrl: String?, repositoryName: String) {
+    setUrl("$artifactoryUrl/$repositoryName")
+    isAllowInsecureProtocol = true
+}
