+++ build-logic/settings.gradle.kts
@@ -1,6 +1,3 @@
-enableFeaturePreview("VERSION_CATALOGS")
-enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
-
 rootProject.name = "build-logic"
 
 include("kotlin")
@@ -10,86 +7,10 @@ include("checks")
 include("publication")
 
 pluginManagement {
-    val artifactoryUrl: String? by settings
-
-    fun MavenArtifactRepository.artifactoryUrl(repositoryName: String) {
-        setUrl("$artifactoryUrl/$repositoryName")
-        isAllowInsecureProtocol = true
-    }
-
-    fun MavenArtifactRepository.setUrlOrProxy(repositoryName: String, originalRepo: String) {
-        if (artifactoryUrl.isNullOrBlank()) {
-            name = repositoryName
-            setUrl(originalRepo)
-        } else {
-            name = "Proxy for $repositoryName: $originalRepo"
-            artifactoryUrl(repositoryName)
-        }
-    }
-
-    repositories {
-        maven {
-            setUrlOrProxy("gradle-plugins", "https://plugins.gradle.org/m2/")
-        }
-    }
+    includeBuild("../build-logic-settings")
 }
 
-val artifactoryUrl: String? by settings
-
-fun MavenArtifactRepository.artifactoryUrl(repositoryName: String) {
-    setUrl("$artifactoryUrl/$repositoryName")
-    isAllowInsecureProtocol = true
-}
-
-fun MavenArtifactRepository.setUrlOrProxy(repositoryName: String, originalRepo: String) {
-    if (artifactoryUrl.isNullOrBlank()) {
-        name = repositoryName
-        setUrl(originalRepo)
-    } else {
-        name = "Proxy for $repositoryName: $originalRepo"
-        artifactoryUrl(repositoryName)
-    }
-}
-
-@Suppress("UnstableApiUsage")
-dependencyResolutionManagement {
-
-    versionCatalogs {
-        create("libs") {
-            from(files("../gradle/libs.versions.toml"))
-        }
-    }
-
-    repositories {
-        maven {
-            setUrlOrProxy("mavenCentral", "https://repo1.maven.org/maven2")
-        }
-        maven {
-            setUrlOrProxy("gradle-plugins", "https://plugins.gradle.org/m2/")
-        }
-        exclusiveContent {
-            forRepository {
-                maven {
-                    setUrlOrProxy("google-android", "https://dl.google.com/dl/android/maven2/")
-                }
-            }
-            filter {
-                includeModuleByRegex("com\\.android.*", "(?!r8).*")
-                includeModuleByRegex("com\\.google\\.android.*", ".*")
-                includeGroupByRegex("androidx\\..*")
-                includeGroup("com.google.testing.platform")
-            }
-        }
-
-        exclusiveContent {
-            forRepository {
-                maven {
-                    setUrlOrProxy("jcenter", "https://jcenter.bintray.com")
-                }
-            }
-            filter {
-                includeGroup("org.jetbrains.trove4j")
-            }
-        }
-    }
+plugins {
+    id("convention-plugins")
+    id("convention-dependencies")
 }
