+++ samples/settings.gradle.kts
@@ -1,120 +1,15 @@
-enableFeaturePreview("VERSION_CATALOGS")
-enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
-
 includeBuild("../build-logic")
 includeBuild("../subprojects")
 
 include(":test-runner")
 
 pluginManagement {
-
-    val artifactoryUrl: String? by settings
-
-    fun MavenArtifactRepository.artifactoryUrl(repositoryName: String) {
-        setUrl("$artifactoryUrl/$repositoryName")
-        isAllowInsecureProtocol = true
-    }
-
-    fun MavenArtifactRepository.setUrlOrProxy(repositoryName: String, originalRepo: String) {
-        if (artifactoryUrl.isNullOrBlank()) {
-            name = repositoryName
-            setUrl(originalRepo)
-        } else {
-            name = "Proxy for $repositoryName: $originalRepo"
-            artifactoryUrl(repositoryName)
-        }
-    }
-
-    @Suppress("UnstableApiUsage")
-    repositories {
-        maven {
-            setUrlOrProxy("mavenCentral", "https://repo1.maven.org/maven2")
-        }
-        exclusiveContent {
-            forRepository {
-                maven {
-                    setUrlOrProxy("google-android", "https://dl.google.com/dl/android/maven2/")
-                }
-            }
-            filter {
-                includeGroupByRegex("com\\.android.*")
-                includeGroupByRegex("androidx.*")
-                includeGroup("com.google.testing.platform")
-            }
-        }
-        maven {
-            setUrlOrProxy("jcenter", "https://jcenter.bintray.com")
-        }
-        maven {
-            setUrlOrProxy("gradle-plugins", "https://plugins.gradle.org/m2/")
-        }
-    }
+    includeBuild("../build-logic-settings")
 }
 
-val artifactoryUrl: String? by settings
-
-fun MavenArtifactRepository.artifactoryUrl(repositoryName: String) {
-    setUrl("$artifactoryUrl/$repositoryName")
-    isAllowInsecureProtocol = true
-}
-
-fun MavenArtifactRepository.setUrlOrProxy(repositoryName: String, originalRepo: String) {
-    if (artifactoryUrl.isNullOrBlank()) {
-        name = repositoryName
-        setUrl(originalRepo)
-    } else {
-        name = "Proxy for $repositoryName: $originalRepo"
-        artifactoryUrl(repositoryName)
-    }
-}
-
-@Suppress("UnstableApiUsage")
-dependencyResolutionManagement {
-
-    versionCatalogs {
-        create("libs") {
-            from(files("../gradle/libs.versions.toml"))
-        }
-    }
-
-    repositories {
-        maven {
-            setUrlOrProxy("mavenCentral", "https://repo1.maven.org/maven2")
-        }
-        exclusiveContent {
-            forRepository {
-                maven {
-                    setUrlOrProxy("google-android", "https://dl.google.com/dl/android/maven2/")
-                }
-            }
-            filter {
-                includeModuleByRegex("com\\.android.*", "(?!r8).*")
-                includeModuleByRegex("com\\.google\\.android.*", ".*")
-                includeGroupByRegex("androidx\\..*")
-                includeGroup("com.google.test.platform")
-            }
-        }
-        exclusiveContent {
-            forRepository {
-                maven {
-                    setUrlOrProxy("KotlinX", "https://kotlin.bintray.com/kotlinx")
-                }
-            }
-            filter {
-                includeModule("org.jetbrains.kotlinx", "kotlinx-html-jvm")
-            }
-        }
-        // not available in mavenCentral
-        exclusiveContent {
-            forRepository {
-                maven {
-                    setUrlOrProxy("jcenter", "https://jcenter.bintray.com")
-                }
-            }
-            filter {
-                includeModule("me.weishu", "free_reflection")
-                includeGroup("com.forkingcode.espresso.contrib")
-            }
-        }
-    }
+plugins {
+    id("convention-plugins")
+    id("convention-cache")
+    id("convention-dependencies")
+    id("convention-scan")
 }
