+++ settings.gradle.kts
@@ -4,5 +4,6 @@ enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 rootProject.name = "avito-android-infra"
 
 includeBuild("build-logic")
+includeBuild("build-logic-settings")
 includeBuild("subprojects")
 includeBuild("samples")
