+++ subprojects/settings.gradle.kts
@@ -131,214 +131,13 @@ include(":logger:elastic-logger")
 
 include(":signer")
 
-@Suppress("UnstableApiUsage")
 pluginManagement {
-
-    val artifactoryUrl: String? by settings
-
-    fun MavenArtifactRepository.artifactoryUrl(repositoryName: String) {
-        setUrl("$artifactoryUrl/$repositoryName")
-        isAllowInsecureProtocol = true
-    }
-
-    fun MavenArtifactRepository.setUrlOrProxy(repositoryName: String, originalRepo: String) {
-        if (artifactoryUrl.isNullOrBlank()) {
-            name = repositoryName
-            setUrl(originalRepo)
-        } else {
-            name = "Proxy for $repositoryName: $originalRepo"
-            artifactoryUrl(repositoryName)
-        }
-    }
-
-    repositories {
-        exclusiveContent {
-            forRepository {
-                mavenLocal()
-            }
-            forRepository {
-                maven {
-                    setUrlOrProxy("mavenCentral", "https://repo1.maven.org/maven2")
-                }
-            }
-            filter {
-                includeModuleByRegex("com\\.avito\\.android", ".*")
-            }
-        }
-        exclusiveContent {
-            forRepository {
-                maven {
-                    setUrlOrProxy("google-android", "https://dl.google.com/dl/android/maven2/")
-                }
-            }
-            filter {
-                includeGroupByRegex("com\\.android.*")
-                includeGroupByRegex("androidx.*")
-                includeGroup("com.google.testing.platform")
-            }
-        }
-        maven {
-            setUrlOrProxy("gradle-plugins", "https://plugins.gradle.org/m2/")
-        }
-    }
-
-    @Suppress("UnstableApiUsage")
-    fun systemProperty(name: String): Provider<String> {
-        return providers.systemProperty(name).forUseAtConfigurationTime()
-    }
-
-    val infraVersion = systemProperty("infraVersion")
-
-    resolutionStrategy {
-        eachPlugin {
-            val pluginId = requested.id.id
-            when {
-                pluginId.startsWith("com.avito.android") ->
-                    useModule("com.avito.android:${pluginId.removePrefix("com.avito.android.")}:${infraVersion.get()}")
-            }
-        }
-    }
-}
-
-val artifactoryUrl: String? by settings
-
-fun MavenArtifactRepository.artifactoryUrl(repositoryName: String) {
-    setUrl("$artifactoryUrl/$repositoryName")
-    isAllowInsecureProtocol = true
-}
-
-fun MavenArtifactRepository.setUrlOrProxy(repositoryName: String, originalRepo: String) {
-    if (artifactoryUrl.isNullOrBlank()) {
-        name = repositoryName
-        setUrl(originalRepo)
-    } else {
-        name = "Proxy for $repositoryName: $originalRepo"
-        artifactoryUrl(repositoryName)
-    }
-}
-
-@Suppress("UnstableApiUsage")
-dependencyResolutionManagement {
-
-    versionCatalogs {
-        create("libs") {
-            from(files("../gradle/libs.versions.toml"))
-        }
-    }
-
-    repositories {
-        maven {
-            setUrlOrProxy("mavenCentral", "https://repo1.maven.org/maven2")
-        }
-
-        // not available in mavenCentral
-        exclusiveContent {
-            forRepository {
-                maven {
-                    setUrlOrProxy("jcenter", "https://jcenter.bintray.com")
-                }
-            }
-            filter {
-                includeGroup("org.jetbrains.trove4j")
-                includeGroup("com.forkingcode.espresso.contrib")
-                includeModule("org.jetbrains.teamcity", "teamcity-rest-client")
-                includeModule("com.fkorotkov", "kubernetes-dsl")
-                includeModule("me.weishu", "free_reflection")
-            }
-        }
-        exclusiveContent {
-            forRepository {
-                maven {
-                    setUrlOrProxy("KotlinX", "https://kotlin.bintray.com/kotlinx")
-                }
-            }
-            filter {
-                includeModule("org.jetbrains.kotlinx", "kotlinx-html-jvm")
-            }
-        }
-        exclusiveContent {
-            forRepository {
-                maven {
-                    setUrlOrProxy("google-android", "https://dl.google.com/dl/android/maven2/")
-                }
-            }
-            filter {
-                includeModuleByRegex("com\\.android.*", "(?!r8).*")
-                includeModuleByRegex("com\\.google\\.android.*", ".*")
-                includeGroupByRegex("androidx\\..*")
-                includeGroup("com.google.testing.platform")
-            }
-        }
-        exclusiveContent {
-            forRepository {
-                maven {
-                    setUrlOrProxy("r8-releases", "https://storage.googleapis.com/r8-releases/raw")
-                }
-            }
-            filter {
-                includeModule("com.android.tools", "r8")
-            }
-        }
-    }
-}
-
-@Suppress("UnstableApiUsage")
-val avitoGithubRemoteCacheHost: Provider<String> = settings.providers
-    .environmentVariable("GRADLE_CACHE_NODE_HOST")
-    .forUseAtConfigurationTime()
-
-val avitoGithubRemoteCachePush: String =
-    extra.properties.getOrDefault("avitoGithub.gradle.buildCache.remote.push", "false").toString()
-
-/**
- * Included builds will inherit this cache config
- * https://docs.gradle.org/current/userguide/build_cache.html#sec:build_cache_composite
- *
- * TODO it is not working as expected with --project-dir subprojects
- *  problem is that's how IDE runs tasks from included builds by default
- *  enabled only for "subprojects" for now
- */
-buildCache {
-    remote<HttpBuildCache> {
-        setUrl("http://${avitoGithubRemoteCacheHost.orNull}/cache/")
-        isEnabled = avitoGithubRemoteCacheHost.orNull != null
-        isPush = avitoGithubRemoteCachePush.toBoolean()
-        isAllowUntrustedServer = true
-        isAllowInsecureProtocol = true
-    }
+    includeBuild("../build-logic-settings")
 }
 
 plugins {
-    id("com.gradle.enterprise") version "3.6.1"
-}
-
-val isCI = booleanProperty("ci", false)
-val buildId = stringProperty("teamcityBuildId", nullIfBlank = true)
-
-gradleEnterprise {
-    buildScan {
-        termsOfServiceUrl = "https://gradle.com/terms-of-service"
-        termsOfServiceAgree = "yes"
-        // Lost scans due to upload interruptions after build finishes
-        isUploadInBackground = false
-        publishAlwaysIf(isCI)
-        if (buildId != null) value("buildId", buildId)
-    }
-}
-
-fun booleanProperty(name: String, defaultValue: Boolean): Boolean {
-    return if (settings.extra.has(name)) {
-        settings.extra[name]?.toString()?.toBoolean() ?: defaultValue
-    } else {
-        defaultValue
-    }
-}
-
-fun stringProperty(name: String, nullIfBlank: Boolean = false): String? {
-    return if (settings.extra.has(name)) {
-        val string = settings.extra[name]?.toString()
-        if (nullIfBlank && string.isNullOrBlank()) null else string
-    } else {
-        null
-    }
+    id("convention-plugins")
+    id("convention-scan")
+    id("convention-cache")
+    id("convention-dependencies")
 }
